Refactor fake filesystem into libs
